pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eleven-audiobooks"
version = "0.1.0"
description = "A Python application that automates the process of converting PDF books into audiobooks"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "sparesparrow", email = "sparesparrow@example.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "aiohttp>=3.8.0",
    "pymongo>=4.0.0",
    "pydantic>=2.0.0",
    "PyPDF2>=3.0.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.3.0",
    "elevenlabs>=0.2.0",
    "deepl>=1.15.0",
    "pydub>=0.25.1",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/sparesparrow/eleven-audiobooks"
Repository = "https://github.com/sparesparrow/eleven-audiobooks.git"
"Bug Tracker" = "https://github.com/sparesparrow/eleven-audiobooks/issues"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=eleven_audiobooks --cov-report=xml --cov-report=term-missing"
testpaths = [
    "tests",
]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = []

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true 